// image-service/app/types/image.types.ts
export type imagemetadata = {
    name: string;
    versionInfo: {
        id: string;
        name: string
    },
    filepath: string;
    fileType: string;
    dimensions: {
        width: number;
        height: number
    },
    thumbnailUrl: string
}
export type imagetransformoptions = {
    "transformations": {
        "resize"?: {
            "width"?: number;
            "height"?: number
        },
        "crop"?: {
            "width"?: number;
            "height"?: number;
            "x": number;
            "y": number
        },
        "rotate"?: number;
        "format"?: string;
        "filters"?: {
            "grayscale": boolean;
            "sepia": boolean
        }
    }
}